Split estimate_gaze into non-ROS (base) and ROS part
