+++ rt_gene/scripts/estimate_gaze.py
@@ -10,41 +10,30 @@ Licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 Interna
 
 from __future__ import print_function, division, absolute_import
 
-import numpy as np
 import os
-import cv2
+import numpy as np
 from tqdm import tqdm
 
+import rospkg
 import rospy
 from sensor_msgs.msg import Image
 from cv_bridge import CvBridge
-import rospkg
-from keras.models import load_model
-from keras.backend.tensorflow_backend import set_session
 
 # noinspection PyUnresolvedReferences
 import rt_gene.gaze_tools as gaze_tools
 from tf import TransformBroadcaster, TransformListener
 import tf.transformations
-import tensorflow
 import collections
 
 from rt_gene.subject_ros_bridge import SubjectListBridge
 from rt_gene.msg import MSG_SubjectImagesList
-from rt_gene.gaze_tools import angle_loss, accuracy_angle
+
+from rt_gene.estimate_gaze_base import GazeEstimatorBase
 
 
-class GazeEstimator(object):
-    """This class encapsulates a deep neural network for gaze estimation.
-
-    It retrieves two image streams, one containing the left eye and another containing the right eye.
-    It synchronizes these two images with the estimated head pose.
-    The images are then converted in a suitable format, and a forward pass (one per eye) of the deep neural network
-    results in the estimated gaze for this frame. The estimated gaze is then published in the (theta, phi) notation.
-    Additionally, two images with the gaze overlaid on the eye images are published."""
-
-    def __init__(self):
-        tqdm.write("PyTorch using {} threads.".format(os.environ["OMP_NUM_THREADS"]))
+class GazeEstimatorROS(GazeEstimatorBase):
+    def __init__(self, device_id_gaze, model_files):
+        super(GazeEstimatorROS, self).__init__(device_id_gaze, model_files)
         self.bridge = CvBridge()
         self.subjects_bridge = SubjectListBridge()
 
@@ -52,75 +41,15 @@ class GazeEstimator(object):
         self.tf_listener = TransformListener()
 
         self.tf_prefix = rospy.get_param("~tf_prefix", "gaze")
-
+        self.headpose_frame = self.tf_prefix + "/head_pose_estimated"
         self.rgb_frame_id_ros = rospy.get_param("~rgb_frame_id_ros", "/kinect2_nonrotated_link")
 
-        self.headpose_frame = self.tf_prefix + "/head_pose_estimated"
-        self.device_id_gazeestimation = rospy.get_param("~device_id_gazeestimation", default="/gpu:0")
-        with tensorflow.device(self.device_id_gazeestimation):
-            config = tensorflow.ConfigProto(inter_op_parallelism_threads=1,
-                                            intra_op_parallelism_threads=1)
-            if "gpu" in self.device_id_gazeestimation:
-                config.gpu_options.allow_growth = True
-                config.gpu_options.per_process_gpu_memory_fraction = 0.3
-            config.log_device_placement = False
-            self.sess = tensorflow.Session(config=config)
-            set_session(self.sess)
-
-        model_files = rospy.get_param("~model_files")
-
-        self.models = []
-        for model_file in model_files:
-            tqdm.write('Load model ' + model_file)
-            model = load_model(os.path.join(rospkg.RosPack().get_path('rt_gene'), model_file),
-                               custom_objects={'accuracy_angle': accuracy_angle, 'angle_loss': angle_loss})
-            # noinspection PyProtectedMember
-            model._make_predict_function()  # have to initialize before threading
-            self.models.append(model)
-        tqdm.write('Loaded ' + str(len(self.models)) + ' models')
-
-        self.graph = tensorflow.get_default_graph()
-        
         self.image_subscriber = rospy.Subscriber('/subjects/images', MSG_SubjectImagesList, self.image_callback, queue_size=1, buff_size=10000000)
         self.subjects_gaze_img = rospy.Publisher('/subjects/gazeimages', Image, queue_size=3)
-       
+
         self.average_weights = np.array([0.1, 0.125, 0.175, 0.2, 0.4])
         self.gaze_buffer_c = {}
         self.time_last = rospy.Time.now()
-
-    def __del__(self):
-        if self.sess is not None:
-            self.sess.close()
-
-    @staticmethod
-    def estimate_gaze_twoeyes_standalone(inference_input_left, inference_input_right, headpose, graph, models):
-        inference_headpose = headpose.reshape(1, 2)
-        with graph.as_default():
-            predictions = []
-            for model in models:
-                predictions.append(model.predict({'img_input_L': inference_input_left,
-                                                  'img_input_R': inference_input_right,
-                                                  'headpose_input': inference_headpose})[0])
-            mean_prediction = np.mean(np.array(predictions), axis=0)
-            if len(models) == 1:  # only apply offset for single model, not for ensemble models
-                mean_prediction[1] += 0.11
-            return mean_prediction
-
-    def estimate_gaze_twoeyes(self, inference_input_left, inference_input_right, headpose):
-        return GazeEstimator.estimate_gaze_twoeyes_standalone(inference_input_left, inference_input_right, headpose, self.graph, self.models)
-
-    @staticmethod
-    def visualize_eye_result(eye_image, est_gaze):
-        """Here, we take the original eye eye_image and overlay the estimated gaze."""
-        output_image = np.copy(eye_image)
-
-        center_x = output_image.shape[1] / 2
-        center_y = output_image.shape[0] / 2
-
-        endpoint_x, endpoint_y = gaze_tools.get_endpoint(est_gaze[0], est_gaze[1], center_x, center_y, 50)
-
-        cv2.line(output_image, (int(center_x), int(center_y)), (int(endpoint_x), int(endpoint_y)), (255, 0, 0))
-        return output_image
 
     def publish_image(self, image, image_publisher, timestamp):
         """This image publishes the `image` to the `image_publisher` with the given `timestamp`."""
@@ -128,20 +57,7 @@ class GazeEstimator(object):
         image_ros.header.stamp = timestamp
         image_publisher.publish(image_ros)
 
-    @staticmethod
-    def input_from_image(cv_image):
-        """This method converts an eye_img_msg provided by the landmark estimator, and converts it to a format
-        suitable for the gaze network."""
-        currimg = cv_image.reshape(36, 60, 3, order='F')
-        currimg = currimg.astype(np.float32)
-        testimg = np.zeros((1, 36, 60, 3))
-        testimg[0, :, :, 0] = currimg[:, :, 0] - 103.939
-        testimg[0, :, :, 1] = currimg[:, :, 1] - 116.779
-        testimg[0, :, :, 2] = currimg[:, :, 2] - 123.68
-        return testimg
-
-    def compute_eye_gaze_estimation(self, subject_id, timestamp,
-                                    input_r,         input_l):
+    def compute_eye_gaze_estimation(self, subject_id, timestamp, input_r, input_l):
         """
         subject_id : integer,  id of the subject
         input_x    : cv_image, input image of x eye
@@ -164,13 +80,13 @@ class GazeEstimator(object):
 
             est_gaze_c = self.estimate_gaze_twoeyes(input_l, input_r, np.array([theta_head, phi_head]))
 
-            self.gaze_buffer_c[subject_id].append(est_gaze_c)           
-            
+            self.gaze_buffer_c[subject_id].append(est_gaze_c)
+
             if len(self.average_weights) == len(self.gaze_buffer_c[subject_id]):
                 est_gaze_c_med = np.average(np.array(self.gaze_buffer_c[subject_id]), axis=0, weights=self.average_weights)
                 self.publish_gaze(est_gaze_c_med, timestamp, subject_id)
                 time_total = (rospy.Time.now() - timestamp).to_sec()
-                tqdm.write('est_gaze_c: {gaze} (fps: {fps:.1f}, latency: {time:.2f}s)'.format(gaze=est_gaze_c_med, fps=1./(rospy.Time.now() - self.time_last).to_sec(), time=time_total))
+                tqdm.write('est_gaze_c: {gaze} (fps: {fps:.1f}, latency: {time:.2f}s)'.format(gaze=est_gaze_c_med, fps=1. / (rospy.Time.now() - self.time_last).to_sec(), time=time_total))
                 self.time_last = rospy.Time.now()
                 return est_gaze_c_med
 
@@ -194,11 +110,11 @@ class GazeEstimator(object):
         for subject_id, s in subjects_dict.items():
             if subject_id not in self.gaze_buffer_c.keys():
                 self.gaze_buffer_c[subject_id] = collections.deque(maxlen=5)
-                
+
             input_r = self.input_from_image(s.right)
             input_l = self.input_from_image(s.left)
             gaze_est = self.compute_eye_gaze_estimation(subject_id, timestamp, input_r, input_l)
-            
+
             if gaze_est is not None:
                 r_gaze_img = self.visualize_eye_result(s.right, gaze_est)
                 l_gaze_img = self.visualize_eye_result(s.left, gaze_est)
@@ -219,16 +135,14 @@ class GazeEstimator(object):
         euler_angle_gaze = gaze_tools.get_euler_from_phi_theta(phi_gaze, theta_gaze)
         quaternion_gaze = tf.transformations.quaternion_from_euler(*euler_angle_gaze)
         self.tf_broadcaster.sendTransform((0, 0, 0.05),  # publish it 5cm above the head pose's origin (nose tip)
-                                          quaternion_gaze,
-                                          msg_stamp,
-                                          self.tf_prefix + "/world_gaze" + str(subject_id),
-                                          self.headpose_frame + str(subject_id))
+                                          quaternion_gaze, msg_stamp, self.tf_prefix + "/world_gaze" + str(subject_id), self.headpose_frame + str(subject_id))
 
 
 if __name__ == '__main__':
     try:
         rospy.init_node('estimate_gaze')
-        gaze_estimator = GazeEstimator()
+        gaze_estimator = GazeEstimatorROS(rospy.get_param("~device_id_gazeestimation", default="/gpu:0"),
+                                          [os.path.join(rospkg.RosPack().get_path('rt_gene'), model_file) for model_file in rospy.get_param("~model_files")])
         rospy.spin()
     except rospy.exceptions.ROSInterruptException:
         print("See ya")
